+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.66.0] - 2022-10-10
+### Added
+- grpc: add TLS support for the outbound through yaml config or dial option.
+
 ## [1.65.0] - 2022-09-26
 ### Added
 - protoc-gen-yarpc-go-v2: Added new yarpc proto plugin `protoc-gen-yarpc-go-v2`. `protoc-gen-yarpc-go-v2`
@@ -1441,6 +1445,7 @@ This release requires regeneration of ThriftRW code.
 ## 0.1.0 - 2016-08-31
 
 - Initial release.
+[1.66.0]: https://github.com/yarpc/yarpc-go/compare/v1.65.0...v1.66.0
 [1.65.0]: https://github.com/yarpc/yarpc-go/compare/v1.64.0...v1.65.0
 [1.64.0]: https://github.com/yarpc/yarpc-go/compare/v1.63.0...v1.64.0
 [1.63.0]: https://github.com/yarpc/yarpc-go/compare/v1.62.0...v1.63.0